Fix critical server failure
That should do it. The error with the hash function was caused because I was reusing the name "hash" in a for loop, and since I switched to Nova, I was forgetting to write const/let for the loop variable, which was done by default in Atom.
The loop was overriding the global hash variable and causing the server to crash the next time it was called as a function.

Of course, all this jank is mostly because I'm too lazy to figure out namespaces for the miscellaneous functions. That will be for another time.