Avoid crash when deleting data files.
With the current setup, the in-memory active sessions index is not aware of file deletions, meaning that the beacon receiver throws if the session file does not exist.
This is fixed for now with a try...catch which outputs the error to the console.

Do note that right now, all active sessions are closed if the server restarts. This should not be a problem in production, but of course with the dev environment server restarts happen all the time so it did amount to a substantial inaccuracy.