Instored error detection for freak incident
A JSON session got corrupted on file, probably by the server itself. It might just have been a power failure so I'm not spending more time on this now. However, if it happens again, the next time the file is accessed, it will clearly be logged and easy to identity.
In this case, it was safe to just delete it. Don't know what could be done if it wasn't though.